Extract shared ANSI escape handling to ANSIHelpers.h
Move common ANSI escape sequence handling code to a shared header:
- isEscapeCharacter() - check for ANSI escape introducers
- findEscapeCharacter() - SIMD-optimized escape sequence search
- consumeANSI() - state machine to consume ANSI sequences

This eliminates ~170 lines of duplicate code between wrapAnsi.cpp
and stripANSI.cpp.

Claude-Generated-By: Claude Code (cli/claude-opus-4-5=100%)
Claude-Steers: 1
Claude-Permission-Prompts: 0
Claude-Escapes: 0
Claude-Plan:
<claude-plan>
# wrapAnsi関数の実装計画

## 概要
BunにNPMの`wrap-ansi`ライブラリと互換性のある`Bun.wrapAnsi()`関数を追加する。

## APIシグネチャ
```typescript
Bun.wrapAnsi(string: string, columns: number, options?: WrapAnsiOptions): string

interface WrapAnsiOptions {
  hard?: boolean;            // default: false - 長い単語を強制分割
  wordWrap?: boolean;        // default: true - 単語境界で折り返し
  trim?: boolean;            // default: true - 行の先頭・末尾の空白を削除
  ambiguousIsNarrow?: boolean; // default: true - 曖昧幅文字をnarrow(幅1)として扱う
}
```

## 実装アプローチ
`stringWidth`と同様にZig + bindgenシステムを使用。既存のANSI/文字幅計算コードを再利用。

## 変更するファイル

### 1. バインディング定義
**ファイル**: `src/bun.js/api/BunObject.bind.ts`

```typescript
export const WrapAnsiOptions = t.dictionary({
  hard: t.boolean.default(false),
  wordWrap: t.boolean.default(true),
  trim: t.boolean.default(true),
  ambiguousIsNarrow: t.boolean.default(true),
});

export const wrapAnsi = fn({
  args: {
    global: t.globalObject,
    str: t.DOMString.default(""),
    columns: t.usize,
    opts: WrapAnsiOptions.default({}),
  },
  ret: t.any, // JSValue (String)
});
```

### 2. Zig実装
**ファイル**: `src/bun.js/api/BunObject.zig`

エントリーポイント関数を追加。コア実装を呼び出す。

### 3. コア実装（新規）
**ファイル**: `src/string/immutable/wrap_ansi.zig`

主要な処理：
1. 改行で分割し、各行を処理
2. スペースで単語に分割し、表示幅を計算（ANSI除外）
3. 列幅に基づいて折り返し
4. ANSIエスケープコードの追跡（SGR、OSC 8ハイパーリンク）
5. 行跨ぎでスタイルを閉じて再開

既存コードの再利用：
- `src/string/immutable/visible.zig` - 表示幅計算
- ANSI CSI/OSCパースロジック

### 4. テスト
**ファイル**: `test/js/bun/util/wrapAnsi.test.ts`

NPMライブラリとの比較テスト：
- 基本的な折り返し
- ANSIカラー付き文字列
- hard/wordWrap/trimオプション
- 全角文字、絵文字、サロゲートペア
- ハイパーリンク

### 5. 型定義
**ファイル**: `packages/bun-types/bun.d.ts`

TypeScript型定義を追加。

## 実装ステップ

### Step 1: bindgen定義追加
`BunObject.bind.ts`にオプションと関数シグネチャを追加。

### Step 2: コア実装
`wrap_ansi.zig`に以下を実装：
- `wrapAnsi()` - メイン関数
- `wrapLine()` - 1行の処理
- `wrapWord()` - 長い単語の分割（hardモード）
- ANSIスタイル追跡構造体
- SGRコードの開始→終了マッピング

### Step 3: BunObjectへの統合
`BunObject.zig`にエントリーポイント関数を追加。

### Step 4: テスト作成
NPMとの互換性テストを作成。

### Step 5: 型定義追加
`bun.d.ts`にインターフェースを追加。

## SGRコードマッピング
行跨ぎでスタイルを復元するために必要：

| 開始コード | 終了コード | 説明 |
|-----------|-----------|------|
| 1 | 22 | bold |
| 2 | 22 | dim |
| 3 | 23 | italic |
| 4 | 24 | underline |
| 7 | 27 | inverse |
| 8 | 28 | hidden |
| 9 | 29 | strikethrough |
| 30-37 | 39 | foreground color |
| 40-47 | 49 | background color |
| 90-97 | 39 | bright foreground |
| 100-107 | 49 | bright background |

## 検証方法

```bash
# テスト実行
bun bd test test/js/bun/util/wrapAnsi.test.ts

# システムBunでは失敗することを確認
USE_SYSTEM_BUN=1 bun test test/js/bun/util/wrapAnsi.test.ts

# 手動テスト
bun bd -e "console.log(Bun.wrapAnsi('hello world', 5))"
```

## 注意事項
- 全角文字は幅2としてカウント
- ANSIエスケープは幅0
- Unicode正規化（normalize()）を適用
- `\r\n`を`\n`に正規化
</claude-plan>
