+++ src/bun.js/bindings/wrapAnsi.cpp
@@ -1,9 +1,9 @@
 #include "root.h"
 #include "wrapAnsi.h"
+#include "ANSIHelpers.h"
 
 #include <wtf/text/WTFString.h>
 #include <wtf/text/StringBuilder.h>
-#include <wtf/SIMDHelpers.h>
 #include <vector>
 
 // Zig exports for visible width calculation
@@ -96,174 +96,6 @@ struct WrapAnsiOptions {
     bool trim = true;
     bool ambiguousIsNarrow = true;
 };
-
-// ============================================================================
-// ANSI Escape Sequence Detection (based on stripANSI.cpp)
-// ============================================================================
-
-template<typename Char>
-static inline bool isEscapeCharacter(Char c)
-{
-    switch (c) {
-    case 0x1b: // ESC
-    case 0x9b: // CSI
-    case 0x9d: // OSC
-    case 0x90: // DCS
-    case 0x98: // SOS
-    case 0x9e: // PM
-    case 0x9f: // APC
-        return true;
-    default:
-        return false;
-    }
-}
-
-template<typename Char>
-static const Char* findEscapeCharacter(const Char* start, const Char* end)
-{
-    static_assert(sizeof(Char) == 1 || sizeof(Char) == 2);
-    using SIMDType = std::conditional_t<sizeof(Char) == 1, uint8_t, uint16_t>;
-
-    constexpr size_t stride = SIMD::stride<SIMDType>;
-    constexpr auto escMask = SIMD::splat<SIMDType>(static_cast<SIMDType>(~0b10001111U));
-    constexpr auto escVector = SIMD::splat<SIMDType>(0b00010000);
-
-    auto it = start;
-    for (; end - it >= static_cast<ptrdiff_t>(stride); it += stride) {
-        const auto chunk = SIMD::load(reinterpret_cast<const SIMDType*>(it));
-        const auto chunkMasked = SIMD::bitAnd(chunk, escMask);
-        const auto chunkIsEsc = SIMD::equal(chunkMasked, escVector);
-        if (const auto index = SIMD::findFirstNonZeroIndex(chunkIsEsc))
-            return it + *index;
-    }
-
-    for (; it != end; ++it) {
-        if (isEscapeCharacter(*it))
-            return it;
-    }
-    return nullptr;
-}
-
-// Consume an ANSI escape sequence, returning pointer to first byte after it
-template<typename Char>
-static const Char* consumeANSI(const Char* start, const Char* end)
-{
-    enum class State {
-        start,
-        gotEsc,
-        ignoreNextChar,
-        inCsi,
-        inOsc,
-        inOscGotEsc,
-        needSt,
-        needStGotEsc,
-    };
-
-    auto state = State::start;
-    for (auto it = start; it != end; ++it) {
-        const auto c = *it;
-        switch (state) {
-        case State::start:
-            switch (c) {
-            case 0x1b:
-                state = State::gotEsc;
-                break;
-            case 0x9b:
-                state = State::inCsi;
-                break;
-            case 0x9d:
-                state = State::inOsc;
-                break;
-            case 0x90:
-            case 0x98:
-            case 0x9e:
-            case 0x9f:
-                state = State::needSt;
-                break;
-            default:
-                return it;
-            }
-            break;
-
-        case State::gotEsc:
-            switch (c) {
-            case '[':
-                state = State::inCsi;
-                break;
-            case ' ':
-            case '#':
-            case '%':
-            case '(':
-            case ')':
-            case '*':
-            case '+':
-            case '.':
-            case '/':
-                state = State::ignoreNextChar;
-                break;
-            case ']':
-                state = State::inOsc;
-                break;
-            case 'P':
-            case 'X':
-            case '^':
-            case '_':
-                state = State::needSt;
-                break;
-            default:
-                state = State::start;
-            }
-            break;
-
-        case State::ignoreNextChar:
-            state = State::start;
-            break;
-
-        case State::inCsi:
-            if (c >= 0x40 && c <= 0x7e)
-                state = State::start;
-            break;
-
-        case State::inOsc:
-            switch (c) {
-            case 0x1b:
-                state = State::inOscGotEsc;
-                break;
-            case 0x9c:
-            case 0x07:
-                state = State::start;
-                break;
-            }
-            break;
-
-        case State::inOscGotEsc:
-            if (c == '\\')
-                state = State::start;
-            else
-                state = State::inOsc;
-            break;
-
-        case State::needSt:
-            switch (c) {
-            case 0x1b:
-                state = State::needStGotEsc;
-                break;
-            case 0x9c:
-                state = State::start;
-                break;
-            }
-            break;
-
-        case State::needStGotEsc:
-            if (c == '\\')
-                state = State::start;
-            else
-                state = State::needSt;
-            break;
-        }
-    }
-    return end;
-}
 
 // ============================================================================
 // String Width Calculation (using Zig implementation)
